+++ internal/ib/sysfs/render_test.go
@@ -1,7 +1,7 @@
 // Copyright 2026 NVIDIA CORPORATION
 // SPDX-License-Identifier: Apache-2.0
 
-package render
+package sysfs
 
 import (
 	"os"
